dev-requirements.txt
-i https://pypi.org/simple

black==22.12.0

click==8.1.3

colorama==0.4.6 ; platform_system == 'Windows'

flake8==6.0.0

mccabe==0.7.0 ; python_version >= '3.6'

mypy-extensions==0.4.3

pathspec==0.10.3 ; python_version >= '3.7'

platformdirs==2.6.2 ; python_version >= '3.7'

pycodestyle==2.10.0 ; python_version >= '3.6'

pyflakes==3.0.1 ; python_version >= '3.6'

tomli==2.0.1 ; python_full_version < '3.11.0a7'

beautifulsoup4==4.11.1 ; python_full_version >= '3.6.0'

bs4==0.0.1

cachetools==5.2.1 ; python_version ~= '3.7'

certifi==2022.12.7 ; python_version >= '3.6'

charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'

copier==7.0.1

dunamai==1.15.0 ; python_version >= '3.5' and python_version < '4.0'

et-xmlfile==1.1.0 ; python_version >= '3.6'

google-api-core==2.11.0 ; python_version >= '3.7'

google-api-python-client==2.72.0

google-auth==2.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'

google-auth-httplib2==0.1.0

google-auth-oauthlib==0.8.0

googleapis-common-protos==1.58.0 ; python_version >= '3.7'

httplib2==0.21.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

idna==3.4 ; python_version >= '3.5'

iteration-utilities==0.11.0 ; python_version >= '3.5'

jinja2==3.1.2

jinja2-ansible-filters==1.3.2

markupsafe==2.1.1 ; python_version >= '3.7'

oauthlib==3.2.2 ; python_version >= '3.6'

openpyxl==3.0.10

packaging==23.0 ; python_version >= '3.7'

plumbum==1.8.1 ; python_version >= '3.6'

prompt-toolkit==3.0.36 ; python_full_version >= '3.6.2'

protobuf==4.21.12 ; python_version >= '3.7'

pyasn1==0.4.8

pyasn1-modules==0.2.8

pydantic==1.10.4 ; python_version >= '3.7'

pygments==2.14.0 ; python_version >= '3.6'

pyparsing==3.0.9 ; python_version >= '3.1'

pypdf2==3.0.1

python-dateutil==2.8.2

python-dotenv==0.21.0

pywin32==305 ; platform_system == 'Windows' and platform_python_implementation != 'PyPy'

pyyaml==6.0 ; python_version >= '3.6'

pyyaml-include==1.3 ; python_version >= '3.5'

questionary==1.10.0

requests==2.28.1 ; python_version >= '3.7' and python_version < '4'

requests-oauthlib==1.3.1

rsa==4.9 ; python_version >= '3.6'

six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

soupsieve==2.3.2.post1 ; python_version >= '3.6'

toml==0.10.2

typing-extensions==4.4.0 ; python_version >= '3.7'

tzdata==2022.7

uritemplate==4.1.1 ; python_version >= '3.6'

urllib3==1.26.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'

wcwidth==0.2.5

